+++ src/templates/stacked_skeleton/app/ui/views/startup_view_model.ts
@@ -11,10 +11,10 @@ export class StartupViewModel extends Base {
 
 import 'package:${YamlHelper.getProjectName()}/src/app/generated/locator/locator.dart';
 import 'package:${YamlHelper.getProjectName()}/src/app/generated/router/router.gr.dart';
-import 'package:stacked/stacked.dart';
+import 'package:${YamlHelper.getProjectName()}/src/ui/global/custom_base_view_model.dart';
 import 'package:stacked_services/stacked_services.dart';
 
-class StartupViewModel extends BaseViewModel {
+class StartupViewModel extends CustomBaseViewModel {
   final NavigationService _navigationService = locator<NavigationService>();
   Future handleStartup() async {
     // Do Some Logic Here